py-mqtt-broker-paho/requirements.txt
paho-python
amqtt
